% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listGeneAF.r
\name{listGeneAF}
\alias{listGeneAF}
\title{list of gene attributes and filters.}
\usage{
listGeneAF(data_set)
}
\arguments{
\item{data_set, }{choose one of '1,2,3,4,5'.'1' = 'btaurus_gene_ensembl','2' = 'ggalluse_gene_ensembl','3' = 'ecaballus_gene_ensembl','4' = 'sscrofa_gene_ensembl','5' = 'oaries_gene_ensembl'.}
}
\value{
result
}
\description{
list of gene attributes and filters.
}
\examples{
listGeneAF(1)
}
