% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listQTLAF.r
\name{listQTLAF}
\alias{listQTLAF}
\title{list of QTL attributes and filters.}
\usage{
listQTLAF()
}
\value{
result
}
\description{
list of QTL attributes and filters.
}
\examples{
listQTLAF()
}
