% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listSNPAttributes.r
\name{listSNPAttributes}
\alias{listSNPAttributes}
\title{list of snp attributes and filters.}
\usage{
listSNPAttributes(data_set)
}
\arguments{
\item{data_set, }{choose one of '1,2,3,4,5'.'1' = 'btaurus_snp',2' = 'ggallus_snp','3' = 'ecaballus_snp',4' = 'sscrofa_snp','5' = 'oaries_s}
}
\value{
result
}
\description{
list of snp attributes and filters.
}
\examples{
listSNPAttributes(1)
}
